+++ library/build-entry.ts
@@ -114,6 +114,8 @@ import {
   Comment,
   ProgressBar,
   InputBadge,
+  FileUpload,
+  UserGroup,
 } from './components';
 
 export {
@@ -188,6 +190,8 @@ export {
   Animation,
   Comment,
   ProgressBar,
+  FileUpload,
+  UserGroup,
 
   // Component store
   useLoadingStore,