+++ library/components/form/Form.vue
@@ -28,21 +28,8 @@ defineSlots<FormSlots>();
 
 onMounted(() => {
   if (fieldsWrapper.value) {
-    const chilren = fieldsWrapper.value.children;
-    const childCount = fieldsWrapper.value.childElementCount;
     const colCount = props.columnPerRow ?? 1;
-
     fieldsWrapper.value.style.gridTemplateColumns = `repeat(${colCount}, minmax(0, 1fr))`;
-
-    let [rowPos, colPos] = [1, 1];
-    for (const i in Array.from({ length: childCount })) {
-      chilren[i].setAttribute('style', `grid-area: ${rowPos}/${colPos}`);
-
-      if (++colPos > colCount) {
-        colPos = 1;
-        rowPos++;
-      }
-    }
 
     setOuterFieldsWrapperHeight();
   }