+++ presets/wangsvue/avatargroup/index.js
@@ -1,5 +1,7 @@
 export default {
   root: {
-    class: 'flex items-center',
+    'class': 'flex items-center ml-4',
+    'data-wv-name': 'usergroup',
+    'data-wv-section': 'root',
   },
 };